ASoC: soc-core: Prevent warning if no DMI table is present
Many systems do not use ACPI and hence do not provide a DMI table. On
non-ACPI systems a warning, such as the following, is printed on boot.

 WARNING KERN tegra-audio-graph-card sound: ASoC: no DMI vendor name!

The variable 'dmi_available' is not exported and so currently cannot be
used by kernel modules without adding an accessor. However, it is
possible to use the function is_acpi_device_node() to determine if the
sound card is an ACPI device and hence indicate if we expect a DMI table
to be present. Therefore, call is_acpi_device_node() to see if we are
using ACPI and only parse the DMI table if we are booting with ACPI.

    Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
    Link: https://lore.kernel.org/r/20210303115526.419458-1-jonathanh@nvidia.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Jon Hunter <jonathanh@nvidia.com>
Date: Wed Mar 3 11:55:26 2021 +0000

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7de14d581dbed57c2b3c6afffa2c3fdc6955a3cd)
Bugzilla: https://bugzilla.redhat.com/1917388