ASoC: tlv320aic32x4: Increase maximum register in regmap
AIC32X4_REFPOWERUP was added as a register, but the maximum register value
in the regmap and regmap range was not correspondingly increased.  This
caused an error when this register was attempted to be written.

    Fixes: ec96690de82c ("ASoC: tlv320aic32x4: Enable fast charge")
    Cc: stable@vger.kernel.org
    Signed-off-by: Annaliese McDermond <nh6z@nh6z.net>
    Link: https://lore.kernel.org/r/0101017889851cab-ce60cfdb-d88c-43d8-bbd2-7fbf34a0c912-000000@us-west-2.amazonses.com
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Annaliese McDermond <nh6z@nh6z.net>
Date: Wed Mar 31 18:21:45 2021 +0000

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 29654ed8384e9dbaf4cfba689dbcb664a6ab4bb7)
Bugzilla: https://bugzilla.redhat.com/1917388