soundwire: bus: Fix device found flag correctly
found flag is used to indicate SoundWire devices that are
both enumerated on the bus and available in the device list.
However this flag is not reset correctly after one iteration,
This could miss some of the devices that are enumerated on the
bus but not in device list. So reset this correctly to fix this issue!

    Fixes: d52d7a1be02c ("soundwire: Add Slave status handling helpers")
    Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
    Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
    Link: https://lore.kernel.org/r/20210309104816.20350-1-srinivas.kandagatla@linaro.org
    Signed-off-by: Vinod Koul <vkoul@kernel.org>

Author: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Date: Tue Mar 9 10:48:16 2021 +0000

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit f03690f4f6992225d05dbd1171212e5be5a370dd)
Bugzilla: https://bugzilla.redhat.com/1917388