ALSA: hdsp: hardware output loopback
Output loopback is a feature where you can record what you hear.
The HDSP series of the RME interfaces provides this functionality
at the hardware level and this patch exposes controls to enable or
disable it per output (playback) channel.

This probably works on other cards but due to a lack of hardware
it is only tested and enabled for the HDSP9632 card with this patch.

Should this patch be accepted a separate patch will be posted to
https://github.com/alsa-project/alsa-tools/tree/master/hdspmixer
which adds "LPBK" buttons to each output in the playback strip for
the user to be able to control this feature from the user land.
Users from Windows tool TotalMixFX should be familiar with this.

    Signed-off-by: Jasmin Fazlic <superfassl@gmail.com>
    Link: https://lore.kernel.org/r/95cb3117-e85a-51a6-c2ce-bf736e70fc4c@gmail.com
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Jasmin Fazlic <superfassl@gmail.com>
Date: Mon Feb 1 15:28:52 2021 +0100

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit da2a040ee7cfe1dd57d5bec7906cb979c5787a86)
Bugzilla: https://bugzilla.redhat.com/1917388