ASoC: codecs/jz4770: Add DAPM widget to set HP out to cap-less mode
Cap-less mode is useful e.g. if the headphones are used as an antenna
for a FM radio, so that the signal is not altered. For everything else,
we want the cap-couple mode.

    Signed-off-by: Paul Cercueil <paul@crapouillou.net>
    Link: https://lore.kernel.org/r/20201207125338.119397-5-paul@crapouillou.net
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Paul Cercueil <paul@crapouillou.net>
Date: Mon Dec 7 12:53:37 2020 +0000

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit e648e3f1165354f04a4deed1f41152a287b68d59)
Bugzilla: https://bugzilla.redhat.com/1917388