ASoC: jz4740-i2s: Remove manual DMA peripheral ID assignment
All platforms that use the jz4740-i2s driver have been switched to
devicetree for a while now and the assignment of the DMA peripheral ID
is done in the devicetree.

It is no longer necessary to manually assign the peripheral ID in the
driver, so remove that. The DMA driver does not even look at the value
assigned in the driver anymore and always uses the value provided by the
devicetree.

    Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
    Reviewed-by: Paul Cercueil <paul@crapouillou.net>
    Link: https://lore.kernel.org/r/20201023092346.5777-1-lars@metafoo.de
    Signed-off-by: Mark Brown <broonie@kernel.org>

Author: Lars-Peter Clausen <lars@metafoo.de>
Date: Fri Oct 23 11:23:46 2020 +0200

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 81dde99f1ac64e27290990c67439b49ecc1f7c29)
Bugzilla: https://bugzilla.redhat.com/1917388