perf/intel: Remove Perfmon-v4 counter_freezing support
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 3daa96d67274653b7c461b30ef9581d68e905fe1
Author: Peter Zijlstra <peterz@infradead.org>
Date: Tue Nov 10 16:37:51 2020 +0100

description
===========
Perfmon-v4 counter freezing is fundamentally broken; remove this default
disabled code to make sure nobody uses it.

The feature is called Freeze-on-PMI in the SDM, and if it would do that,
there wouldn't actually be a problem, *however* it does something subtly
different. It globally disables the whole PMU when it raises the PMI,
not when the PMI hits.

This means there's a window between the PMI getting raised and the PMI
actually getting served where we loose events and this violates the
perf counter independence. That is, a counting event should not result
in a different event count when there is a sampling event co-scheduled.

This is known to break existing software (RR).

Conflicts:
==========
Not backported 3a4ac121c2cacbf97d49 ("x86/perf: Add hardware performance
events support for Zhaoxin CPU.") caused different context.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>