perf/x86/rapl: Add msr mask support
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit ffb20c2e52e8709b5fc9951e8863e31efb1f2cba
Author: Zhang Rui <rui.zhang@intel.com>
Date: Fri Feb 5 00:18:14 2021 +0800

description
===========
In some cases, when probing a perf MSR, we're probing certain bits of the
MSR instead of the whole register, thus only these bits should be checked.

For example, for RAPL ENERGY_STATUS MSR, only the lower 32 bits represents
the energy counter, and the higher 32bits are reserved.

Introduce a new mask field in struct perf_msr to allow probing certain
bits of a MSR.

This change is transparent to the current perf_msr_probe() users.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>