perf/x86/intel: Set PERF_ATTACH_SCHED_CB for large PEBS and LBR
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit afbef30149587ad46f4780b1e0cc5e219745ce90
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Mon Nov 30 11:38:41 2020 -0800

description
===========
To supply a PID/TID for large PEBS, it requires flushing the PEBS buffer
in a context switch.

For normal LBRs, a context switch can flip the address space and LBR
entries are not tagged with an identifier, we need to wipe the LBR, even
for per-cpu events.

For LBR callstack, save/restore the stack is required during a context
switch.

Set PERF_ATTACH_SCHED_CB for the event with large PEBS & LBR.

Fixes: 9c964efa4330 ("perf/x86/intel: Drain the PEBS buffer during context switches")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>