perf c2c: Refactor node display
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit f3d0a551db13b7cdf8addbe265b4fd7026944d24
Author: Leo Yan <leo.yan@linaro.org>
Date: Thu Jan 14 23:46:45 2021 +0800

description
===========
The macro DISPLAY_HITM() is used to calculate HITM percentage introduced
by every node and it's shown for the node info.

This patch introduces the static function display_metrics() to replace
the macro, and the parameters are refined for passing the metric's
statistic and sum value.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>