perf tools: Add 'ping' control command
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 47fddcb479e73c341fb414e40a89572dacadd360
Author: Jiri Olsa <jolsa@kernel.org>
Date: Sun Dec 27 00:20:38 2020 +0100

description
===========
Add a control 'ping' command to detect if perf is up and its control
interface is operational.

It will be used in following daemon patches to synchronize with record
session - when control interface is up and running, we know that perf
record is monitoring and ready to receive signals.

Example session:

  terminal 1:

    # mkfifo control ack
    # perf record --control=fifo:control,ack

  terminal 2:

    # echo ping > control
    # cat ack
    ack

Signed-off-by: Michael Petlan <mpetlan@redhat.com>