perf vendor events arm64: Add common and uarch event JSON
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit c77669662f1a86bce38dcad5b6da9eac6be35ec3
Author: John Garry <john.garry@huawei.com>
Date: Thu Jan 28 20:00:34 2021 +0800

description
===========
Add a common and microarch JSON, which can be referenced from CPU JSONs.

For now, brief and public description are as event brief event
description from the ARMv8 ARM [0], D7-11.

The list of events is not complete, as not all events will be referenced
yet.

Reference document is at the following:

[0] https://documentation-service.arm.com/static/5fa3bd1eb209f547eebd4141?token=

Signed-off-by: Michael Petlan <mpetlan@redhat.com>