perf tools: Use /proc/<PID>/task/<TID>/status for PERF_RECORD_ event synthesis
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 30626e0844757b7b371353eb693354c4ccc5eb0e
Author: Namhyung Kim <namhyung@kernel.org>
Date: Tue Feb 2 18:01:16 2021 +0900

description
===========
To save memory usage, it needs to reduce the number of entries in the
proc filesystem.  It's using /proc/<PID>/task directory to traverse
threads in the process and then kernel creates /proc/<PID>/task/<TID>
entries.

After that it checks the thread info using the /proc/<TID>/status file
rather than /proc/<PID>/task/<TID>/status.  As far as I can see, they
are the same and contain all the info we need.

Using the latter eliminates the unnecessary /proc/<TID> entry.  This can
be useful especially a large number of threads are used in the system.
In my experiment around 1KB of memory on average was saved for each
thread (which is not a thread group leader).

To do this, pass both pid and tid to perf_event_prepare_comm() if it
knows them.  In case it doesn't know, passing 0 as pid will do the old
way.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>