perf inject jit: Add namespaces support
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 67dec926931448d688efb5fe34f7b5a22470fc0a
Author: Yonatan Goldschmidt <yonatan.goldschmidt@granulate.io>
Date: Thu Nov 5 03:56:04 2020 +0200

description
===========
This patch fixes "perf inject --jit" to properly operate on
namespaced/containerized processes:

* jitdump files are generated by the process, thus they should be
  looked up in its mount NS.

* DSOs of injected MMAP events will later be looked up in the process
  mount NS, so write them into its NS.

* PIDs & TIDs from jitdump events need to be translated to the PID as
  seen by "perf record" before written into MMAP events.

For a process in a different PID NS, the TID & PID given in the jitdump
event are actually ignored; I use the TID & PID of the thread which
mmap()ed the jitdump file. This is simplified and won't do for forks of
the initial process, if they continue using the same jitdump file.
Future patches might improve it.

This was tested by recording a NodeJS process running with
"--perf-prof", inside a Docker container, and by recording another
NodeJS process running in the same namespaces as perf itself, to make
sure it's not broken for non-containerized processes.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>