perf arm64/s390: Fix printf conversion specifier for IP addresses
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 6db59d357e8e720c91963dfd69604d238d2143cd
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Feb 9 10:17:01 2021 -0300

description
===========
We need to use "%#" PRIx64 for u64 values, not "%lx". In arm64's and
s390x cases the compiler doesn't complain, but lets fix this in case
this code gets copied to a 32-bit arch, like with powerpc 32-bit that
got fixed in the previous patch.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>