perf daemon: Add daemon command
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit d450bc501fbdceb9d71663ba8192b72f01001bf1
Author: Jiri Olsa <jolsa@kernel.org>
Date: Mon Feb 8 21:08:45 2021 +0100

description
===========
Add a daemon skeleton with a minimal base (non) functionality, covering
various setup in start command.

Add an initial perf-daemon.txt with basic info.

This is in response to pople asking for the possibility to be able run
record long running sessions on the background.

The patchset that starts with this adds support to configure and run
record sessions on background via new 'perf daemon' command.

This is useful for being able to use perf as a flight recorder that one
can interact with asking for events to be enabled or disabled, added or
removed, etc.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>