tools api fs: Prefer cgroup v1 path in cgroupfs_find_mountpoint()
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 27ab1c1c06529056df498b9647c03167e682b4d9
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Dec 16 18:05:54 2020 +0900

description
===========
The cgroupfs_find_mountpoint() looks up the /proc/mounts file to find
a directory for the given cgroup subsystem.  It keeps both cgroup v1
and v2 path since there's a possibility of the mixed hierarchly.

But we can simply use v1 path if it's found as it will override the v2
hierarchy.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>