perf intel-pt: Fix premature IPC
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 20aa39708a5999b7921b27482a756766272286ac
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Feb 5 19:53:48 2021 +0200

description
===========
The code assumed a change in cycle count means accurate IPC. That is not
correct, for example when sampling both branches and instructions, or at
a FUP packet (which is not CYC-eligible) address. Fix by using an explicit
flag to indicate when IPC can be sampled.

Fixes: 5b1dc0fd1da06 ("perf intel-pt: Add support for samples to contain IPC ratio")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>