perf auxtrace: Automatically group aux-output events
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit d58b3f7e701408ac41b0a8342f26bbf9a5597f05
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Jan 21 16:04:18 2021 +0200

description
===========
aux-output events need to have an AUX area event as the group leader.
However, grouping events does not allow the AUX area event to be given
an address filter because the --filter option must come after the event,
which conflicts with the grouping syntax.

To allow filtering in that case, automatically create a group since that
is the requirement anyway.

Example: (requires Intel Tremont)

  perf record -c 500 -e 'intel_pt//u' --filter 'filter main @ /bin/ls' -e 'cycles/aux-output/pp' ls

Signed-off-by: Michael Petlan <mpetlan@redhat.com>