perf intel-pt: Amend decoder to track the NR flag
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 80a038860b5ca3aa864756ad6bfe1af5ac5839b0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Feb 18 11:57:54 2021 +0200

description
===========
The PIP packet NR (non-root) flag indicates whether or not a virtual
machine is being traced (NR=1 => VM). Add support for tracking its value.

In particular note that the PIP packet (outside of PSB+) will be
associated with a TIP packet from which address the NR value takes
effect. At that point, there is a branch from_ip, to_ip with
corresponding from_nr and to_nr.

In the event of VM-Entry failure, there should still PIP and TIP packets
that can be followed in the same way.

Also note that this assumes that a host VMM is not employing VMX controls
that affect Intel PT, e.g. to hide the host from a guest using Intel PT.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>