perf test: Suppress logs in cs-etm testing
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 46355e3d793c6d67fbfb4b155abd0869b6a7d79b
Author: Leo Yan <leo.yan@linaro.org>
Date: Mon Feb 15 19:59:43 2021 +0800

description
===========
With the option '-v' for the verbose logs, "perf test" outputs tons of
logs for the CoreSight case, the logs are mainly introduced by the
decoding.  And it outputs some trivial info from "perf record" command
and there have debugging info for CPU number and device name when
iterates between ETMs and sinks.

For a neat output format, this patch redirects the output logs to
"/dev/null", thus can avoid to flood logs.  And it removes the redundant
log for CPU number and device name, which have already printed out the
relevant info in the function record_touch_file().

Signed-off-by: Michael Petlan <mpetlan@redhat.com>