perf buildid-cache: Add test for 16-byte build-id
Bugzilla: https://bugzilla.redhat.com/1944685

upstream
========
commit 206236d328ee02b171188dfc961aa0d13943c76e
Author: Nicholas Fraser <nfraser@codeweavers.com>
Date: Tue Feb 16 15:38:08 2021 -0500

description
===========
tests/shell/buildid.sh added an ELF executable with an MD5 build-id to
the perf debug cache but did not check whether the object was printed
by a subsequent call to "perf buildid-cache -l". It was being omitted
from the list.

A previous commit fixed the bug that left it out of the list. This adds
a test for it.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>