ALSA: firewire-lib: fix the context to call snd_pcm_stop_xrun()
In the workqueue to queue wake-up event, isochronous context is not
processed, thus it's useless to check context for the workqueue to switch
status of runtime for PCM substream to XRUN. On the other hand, in
software IRQ context of 1394 OHCI, it's needed.

This commit fixes the bug introduced when tasklet was replaced with
workqueue.

    Cc: <stable@vger.kernel.org>
    Fixes: 2b3d2987d800 ("ALSA: firewire: Replace tasklet with work")
    Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>
    Link: https://lore.kernel.org/r/20210605091054.68866-1-o-takashi@sakamocchi.jp
    Signed-off-by: Takashi Iwai <tiwai@suse.de>

Author: Takashi Sakamoto <o-takashi@sakamocchi.jp>
Date: Sat Jun 5 18:10:54 2021 +0900

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 9981b20a5e3694f4625ab5a1ddc98ce7503f6d12)
Bugzilla: https://bugzilla.redhat.com/1917388