Enable CRYPTO_SHA512 and _SSSE3 so these are available in FIPS mode
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1973106
Upstream Status: RHEL-only
Tested: a manual boot in a FIPS mode, see bz1973106#c11

Currently the CRYPTO_SHA512 is built as a module. This results in a hung
task while booting in a FIPS mode:

[  246.750959] INFO: task modprobe:260 blocked for more than 120 seconds.
[  246.755153]       Not tainted 4.18.0-314.el8.x86_64 #1
[  246.760045] "echo 0 > /proc/sys/kernel/hung_task_timeout_secs" disables this message.
[  246.766918] task:modprobe state:D stack:0 pid:260 ppid:8 flags:0x00000000

Fix this by embedding the CRYPTO_SHA512 code into a kernel, the same way
as we did for CTS, CBC and other algos needed in a FIPS mode.

This is somewhat urgent, as it blocks a FIPS testing for the RHEL-8.5.0
and will become a release blocker after Beta.

Signed-off-by: Vladis Dronov <vdronov@redhat.com>