crypto: drbg - self test for HMAC(SHA-512)
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1973106
Upstream Status: merged into the herbert/cryptodev-2.6.git
Tested: a manual boot in a FIPS mode, see bz1973106#c11

Conflicts: This commits is not in the upstream yet, but should
get there as it is. I'm adding it because it is trivial and
fixes a real issue.

commit 8833272d876eb5320e4c7c31cc455542510dc4c3
Author: Stephan Müller <smueller@chronox.de>
Date:   Thu Jun 24 17:44:35 2021 +0200

    crypto: drbg - self test for HMAC(SHA-512)

    Considering that the HMAC(SHA-512) DRBG is the default DRBG now, a self
    test is to be provided.

    The test vector is obtained from a successful NIST ACVP test run.

    Signed-off-by: Stephan Mueller <smueller@chronox.de>
    Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

Signed-off-by: Vladis Dronov <vdronov@redhat.com>